rename profile.common to profile.serialisation
