+++ profile/src/main/java/com/scottlogic/deg/profile/serialisation/ProfileDeserialiser.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.profile.common;
+package com.scottlogic.deg.profile.serialisation;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
